refactor: move legacy clean up to fixtures
